Uncouple path generation from drunc.py scenario bounds.
Previously, NetworkAgentPath and SidewalkAgentPath checked whether
sampled points lie in the scenario bounds in drunc.py. This feature is
now removed and check is instead done with respect to the constraints
specified in the arguments.

Effectively, whether or not the scenario bounds is involved is delegated
to the caller of the function, instead of the function itself.

This also fixes a previous bug where the ego actor goes out of scenario
bounds - SidewalkAgentPath previously had the scenario bounds check, so
it is impossible to sample any point that lies in both the
out-of-scenario-bounds ego-centric bounds and the scenario bounds. This
led to an infinite loop.